apps/employee_management/serializers.py
from rest_framework import serializers

from apps.employee_management.models import Employee


class EmployeeSerializer(serializers.ModelSerializer):
    class Meta:
        model = Employee
        fields = "__all__"
